requirements.txt
spotipy
requests
beautifulsoup4
python-dotenv
